defined object relationships and fixed some typos in our migrations. added a city to the address table

+++ db/migrate/20190423184756_create_neighborhoods.rb
@@ -2,8 +2,8 @@ class CreateNeighborhoods < ActiveRecord::Migration[5.2]
   def change
     create_table :neighborhoods do |t|
       t.string :name
-      t.double :lat
-      t.double :long
+      t.float :lat
+      t.float :long
 
       t.timestamps
     end
